Fix interactive shell and keyboard handling
- Added real keyboard interrupt handler (IRQ1)
- Fixed shell input/output loop
- Connected keyboard driver to shell
- Added proper boot sequence with user feedback
- Fixed kernel main to actually start interactive shell

+++ GROK/ternarybit-os/kernel/gui.c
@@ -42,7 +42,7 @@ typedef struct {
     uint16_t bpp;
     uint8_t* framebuffer;
     uint32_t pitch;
-    uint8_t mode;
+    uint16_t mode;
 } display_mode_t;
 
 typedef struct {
